pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Workspace project"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.69.0",
    "beautifulsoup4>=4.14.2",
    "claude-code-sdk>=0.0.20",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "langchain>=0.2.1",
    "langchain-openai>=0.3.28",
    "markdownify>=1.2.0",
    "networkx>=3.5",
    "openai>=1.108.1",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=1.0.10",
    "pydantic-settings>=2.10.1",
    "pydot>=4.0.1",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
    "yt-dlp>=2025.9.26",
]

[tool.uv.workspace]
# Add all projects in the workspace
members = []

[tool.uv.sources]
# Example: my-project = { workspace = true }

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "debugpy>=1.8.14",
    "pyright>=1.1.406",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
    "twine>=6.1.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [
    "**/__pycache__",
    ".venv/**",
    "scenarios/*/.venv/**",
    "node_modules/**",
    ".git/**",
    ".ruff_cache/**",
    ".pytest_cache/**",
    "*.egg-info",
    "build/**",
    "dist/**",
    ".data/**",
    "ai_working/**",
]
typeCheckingMode = "basic"
reportMissingImports = false
pythonVersion = "3.11"
